Move code for checking roles and permissions from galley-types to wire-api
