+++ libs/galley-types/galley-types.cabal
@@ -97,7 +97,6 @@ test-suite galley-types-tests
   -- cabal-fmt: expand test
   other-modules:
     Paths_galley_types
-    Test.Galley.Permissions
     Test.Galley.Roundtrip
     Test.Galley.Types
 
@@ -153,13 +152,10 @@ test-suite galley-types-tests
   build-depends:
       aeson
     , base
-    , containers
     , galley-types
     , imports
-    , lens
     , QuickCheck
     , tasty
-    , tasty-hunit
     , tasty-quickcheck
     , wire-api
