+++ libs/galley-types/src/Galley/Types/Teams.hs
@@ -56,95 +56,20 @@ module Galley.Types.Teams
     isTeamMember,
     isTeamOwner,
     canSeePermsOf,
-    rolePermissions,
-    roleHiddenPermissions,
-    permissionsRole,
-    isAdminOrOwner,
-    HiddenPerm (..),
-    IsPerm (..),
   )
 where
 
-import Control.Lens (makeLenses, view, (^.))
+import Control.Lens (makeLenses, view)
 import Data.Aeson
 import Data.Aeson.Types qualified as A
 import Data.Id (UserId)
-import Data.Maybe qualified as Maybe
 import Data.Schema qualified as Schema
 import Data.Set qualified as Set
 import Imports
 import Test.QuickCheck (Arbitrary)
-import Wire.API.Error.Galley
 import Wire.API.Team.Feature
 import Wire.API.Team.Member
 import Wire.API.Team.Permission
-import Wire.API.Team.Role
-
-rolePermissions :: Role -> Permissions
-rolePermissions role = Permissions p p where p = rolePerms role
-
-permissionsRole :: Permissions -> Maybe Role
-permissionsRole (Permissions p p') =
-  if p /= p'
-    then do
-      -- we never did use @p /= p'@ for anything, fingers crossed that it doesn't occur anywhere
-      -- in the wild.  but if it does, this implementation prevents privilege escalation.
-      let p'' = Set.intersection p p'
-       in permissionsRole (Permissions p'' p'')
-    else permsRole p
-  where
-    permsRole :: Set Perm -> Maybe Role
-    permsRole perms =
-      Maybe.listToMaybe
-        [ role
-          | role <- [minBound ..],
-            -- if a there is a role that is strictly less permissive than the perms set that
-            -- we encounter, we downgrade.  this shouldn't happen in real life, but it has
-            -- happened to very old users on a staging environment, where a user (probably)
-            -- was create before the current publicly visible permissions had been stabilized.
-            rolePerms role `Set.isSubsetOf` perms
-        ]
-
-isAdminOrOwner :: Permissions -> Bool
-isAdminOrOwner perms =
-  case permissionsRole perms of
-    Just RoleOwner -> True
-    Just RoleAdmin -> True
-    Just RoleMember -> False
-    Just RoleExternalPartner -> False
-    Nothing -> False
-
--- | Internal function for 'rolePermissions'.  (It works iff the two sets in 'Permissions' are
--- identical for every 'Role', otherwise it'll need to be specialized for the resp. sides.)
-rolePerms :: Role -> Set Perm
-rolePerms RoleOwner =
-  rolePerms RoleAdmin
-    <> Set.fromList
-      [ GetBilling,
-        SetBilling,
-        DeleteTeam
-      ]
-rolePerms RoleAdmin =
-  rolePerms RoleMember
-    <> Set.fromList
-      [ AddTeamMember,
-        RemoveTeamMember,
-        SetTeamData,
-        SetMemberPermissions
-      ]
-rolePerms RoleMember =
-  rolePerms RoleExternalPartner
-    <> Set.fromList
-      [ DeleteConversation,
-        AddRemoveConvMember,
-        ModifyConvName,
-        GetMemberPermissions
-      ]
-rolePerms RoleExternalPartner =
-  Set.fromList
-    [ CreateConversation,
-      GetTeamConversations
-    ]
 
 -- This is the cassandra timestamp of writetime(binding)
 newtype TeamCreationTime = TeamCreationTime
@@ -303,104 +228,6 @@ instance ToJSON FeatureTeamSearchVisibilityAvailability where
 makeLenses ''TeamCreationTime
 makeLenses ''FeatureFlags
 makeLenses ''Defaults
-
--- Note [hidden team roles]
---
--- The problem: the mapping between 'Role' and 'Permissions' is fixed by external contracts:
--- client apps treat permission bit matrices as opaque role identifiers, so if we add new
--- permission flags, things will break there.
---
--- "Hidden" in "HiddenPerm", therefore, refers to a permission hidden from
--- clients, thereby making it internal to the backend.
---
--- The solution: add new permission bits to 'HiddenPerm', 'HiddenPermissions', and make
--- 'hasPermission', 'mayGrantPermission' polymorphic.  Now you can check both for the hidden
--- permission bits and the old ones that we share with the client apps.
-
--- | See Note [hidden team roles]
-data HiddenPerm
-  = ChangeLegalHoldTeamSettings
-  | ChangeLegalHoldUserSettings
-  | ViewLegalHoldUserSettings
-  | ChangeTeamFeature
-  | ChangeTeamSearchVisibility
-  | ViewTeamSearchVisibility
-  | ViewSameTeamEmails
-  | ReadIdp
-  | CreateUpdateDeleteIdp
-  | CreateReadDeleteScimToken
-  | -- | this has its own permission because we're not sure how
-    -- efficient this end-point is.  better not let all team members
-    -- play with it unless we have to.
-    DownloadTeamMembersCsv
-  | ChangeTeamMemberProfiles
-  | SearchContacts
-  deriving (Eq, Ord, Show)
-
--- | See Note [hidden team roles]
-data HiddenPermissions = HiddenPermissions
-  { _hself :: Set HiddenPerm,
-    _hcopy :: Set HiddenPerm
-  }
-  deriving (Eq, Ord, Show)
-
-makeLenses ''HiddenPermissions
-
-roleHiddenPermissions :: Role -> HiddenPermissions
-roleHiddenPermissions role = HiddenPermissions p p
-  where
-    p = roleHiddenPerms role
-    roleHiddenPerms :: Role -> Set HiddenPerm
-    roleHiddenPerms RoleOwner = roleHiddenPerms RoleAdmin
-    roleHiddenPerms RoleAdmin =
-      (roleHiddenPerms RoleMember <>) $
-        Set.fromList
-          [ ChangeLegalHoldTeamSettings,
-            ChangeLegalHoldUserSettings,
-            ChangeTeamSearchVisibility,
-            ChangeTeamFeature,
-            ChangeTeamMemberProfiles,
-            ReadIdp,
-            CreateUpdateDeleteIdp,
-            CreateReadDeleteScimToken,
-            DownloadTeamMembersCsv
-          ]
-    roleHiddenPerms RoleMember =
-      (roleHiddenPerms RoleExternalPartner <>) $
-        Set.fromList
-          [ ViewSameTeamEmails,
-            SearchContacts
-          ]
-    roleHiddenPerms RoleExternalPartner =
-      Set.fromList
-        [ ViewLegalHoldUserSettings,
-          ViewTeamSearchVisibility
-        ]
-
--- | See Note [hidden team roles]
-class IsPerm perm where
-  type PermError (e :: perm) :: GalleyError
-
-  roleHasPerm :: Role -> perm -> Bool
-  roleGrantsPerm :: Role -> perm -> Bool
-  hasPermission :: TeamMember -> perm -> Bool
-  hasPermission tm perm = maybe False (`roleHasPerm` perm) . permissionsRole $ tm ^. permissions
-  mayGrantPermission :: TeamMember -> perm -> Bool
-  mayGrantPermission tm perm = maybe False (`roleGrantsPerm` perm) . permissionsRole $ tm ^. permissions
-
-instance IsPerm Perm where
-  type PermError p = 'MissingPermission ('Just p)
-
-  roleHasPerm r p = p `Set.member` (rolePermissions r ^. self)
-  roleGrantsPerm r p = p `Set.member` (rolePermissions r ^. copy)
-  hasPermission tm p = p `Set.member` (tm ^. permissions . self)
-  mayGrantPermission tm p = p `Set.member` (tm ^. permissions . copy)
-
-instance IsPerm HiddenPerm where
-  type PermError p = OperationDenied
-
-  roleHasPerm r p = p `Set.member` (roleHiddenPermissions r ^. hself)
-  roleGrantsPerm r p = p `Set.member` (roleHiddenPermissions r ^. hcopy)
 
 notTeamMember :: [UserId] -> [TeamMember] -> [UserId]
 notTeamMember uids tmms =
