+++ libs/galley-types/test/unit/Test/Galley/Types.hs
@@ -20,63 +20,16 @@
 
 module Test.Galley.Types where
 
-import Control.Lens
-import Data.Set hiding (drop)
-import Data.Set qualified as Set
 import Galley.Types.Teams
 import Imports
 import Test.Galley.Roundtrip (testRoundTrip)
 import Test.QuickCheck qualified as QC
 import Test.Tasty
-import Test.Tasty.HUnit
 import Test.Tasty.QuickCheck
 import Wire.API.Team.Feature as Public
-import Wire.API.Team.Permission
-import Wire.API.Team.Role
 
 tests :: TestTree
-tests =
-  testGroup
-    "Tests"
-    [ testCase "owner has all permissions" $
-        rolePermissions RoleOwner @=? fullPermissions,
-      testCase "smaller roles (further to the left/top in the type def) are strictly more powerful" $
-        -- we may not want to maintain this property in the future when adding more roles, but for
-        -- now it's true, and it's nice to have that written down somewhere.
-        forM_ [(r1, r2) | r1 <- [minBound ..], r2 <- drop 1 [r1 ..]] $
-          \(r1, r2) -> do
-            assertBool "owner.self" ((rolePermissions r2 ^. self) `isSubsetOf` (rolePermissions r1 ^. self))
-            assertBool "owner.copy" ((rolePermissions r2 ^. copy) `isSubsetOf` (rolePermissions r1 ^. copy)),
-      testRoundTrip @FeatureFlags,
-      testGroup
-        "permissionsRole, rolePermissions"
-        [ testCase "'Role' maps to expected permissions" $ do
-            assertEqual "role type changed" [minBound ..] [RoleOwner, RoleAdmin, RoleMember, RoleExternalPartner]
-            assertEqual "owner" (permissionsRole =<< newPermissions (intToPerms 8191) (intToPerms 8191)) (Just RoleOwner)
-            assertEqual "admin" (permissionsRole =<< newPermissions (intToPerms 5951) (intToPerms 5951)) (Just RoleAdmin)
-            assertEqual "member" (permissionsRole =<< newPermissions (intToPerms 1587) (intToPerms 1587)) (Just RoleMember)
-            assertEqual "external partner" (permissionsRole =<< newPermissions (intToPerms 1025) (intToPerms 1025)) (Just RoleExternalPartner),
-          testCase "Role <-> Permissions roundtrip" $ do
-            assertEqual "admin" (permissionsRole . rolePermissions <$> [minBound ..]) (Just <$> [minBound ..]),
-          testProperty "Random, incoherent 'Permission' values gracefully translate to subsets." $
-            let fakeSort (w, w') = (w `Set.union` w', w')
-             in \(fakeSort -> (w, w')) -> do
-                  let Just perms = newPermissions w w'
-                  case permissionsRole perms of
-                    Just role -> do
-                      let perms' = rolePermissions role
-                      assertEqual "eq" (perms' ^. self) (perms' ^. copy)
-                      assertBool "self" ((perms' ^. self) `Set.isSubsetOf` (perms ^. self))
-                      assertBool "copy" ((perms' ^. copy) `Set.isSubsetOf` (perms ^. copy))
-                    Nothing -> do
-                      let leastPermissions = rolePermissions maxBound
-                      assertBool "no role for perms, but strictly more perms than max role" $
-                        not
-                          ( (leastPermissions ^. self) `Set.isSubsetOf` w
-                              && (leastPermissions ^. copy) `Set.isSubsetOf` w'
-                          )
-        ]
-    ]
+tests = testGroup "Tests" [testRoundTrip @FeatureFlags]
 
 instance Arbitrary FeatureFlags where
   arbitrary =
