+++ services/spar/src/Spar/Intra/Galley.hs
@@ -26,7 +26,6 @@ import Control.Lens
 import Control.Monad.Except
 import Data.ByteString.Conversion
 import Data.Id (TeamId, UserId)
-import Galley.Types.Teams
 import Imports
 import Network.HTTP.Types.Method
 import Spar.Error
