+++ tools/stern/stern.cabal
@@ -83,7 +83,6 @@ library
     , errors                 >=1.4
     , exceptions             >=0.6
     , extended
-    , galley-types           >=0.81.0
     , http-client            >=0.7
     , http-types             >=0.8
     , imports
